fixed IngredientDetail amount validation

+++ app/Http/Requests/IngredientDetailFormRequest.php
@@ -24,7 +24,7 @@ class IngredientDetailFormRequest extends FormRequest
     public function rules()
     {
         return [
-            'amount'                => ['required', 'integer'],
+            'amount'                => ['required', 'numeric'],
             'unit_id'               => ['required', 'integer'],
             'ingredient_id'         => ['required', 'integer'],
             'prep_id'               => ['required', 'integer'],
